requirementes.txt
pandas>=2.2
pyyaml>=6.0
numpy>=1.26
opencv-python>=4.9
Pillow>=10.0
ultralytics>=8.2
pyarrow>=16.0
pytest>=8.0
